pets95/requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==3.1.7
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-import-export==2.5.0
django-js-asset==1.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
static3==0.7.0
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
